% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_and_leave_one_out.R
\name{CIs_for_target_state}
\alias{CIs_for_target_state}
\title{Ancillary function of \code{boot_ebmstate}.}
\usage{
CIs_for_target_state(target_state, probtrans_objects_boot)
}
\arguments{
\item{target_state}{The target state for whose transition probabilties the confidence intervals
are computed.}

\item{probtrans_objects_boot}{A list containing bootstrap estimates of transition probabilities.}
}
\value{
95\% highest density bootstrap confidence intervals for the transition
probabilities into \code{target_state}.
}
\description{
Computes 95\% highest density bootstrap confidence 
intervals for the transition probabilities into \code{target_state}, 
given a list object with boostrap estimates of transition probabilities into multiple states. This 
function is not meant to be called by the user.
}
\details{
Uses function \code{extract_function}.
}
\seealso{
\code{\link{boot_ebmstate}}; \code{\link{extract_function}}.
}
\author{
Rui Costa
}
