% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolute_clockforward}
\alias{convolute_clockforward}
\title{Convolution function for clock-forward models}
\usage{
convolute_clockforward(
  time_vector,
  diff_vector,
  probtrans_vector_1,
  probtrans_vector_2
)
}
\arguments{
\item{time_vector, diff_vector, probtrans_vector_1, probtrans_vector_2}{Numeric vectors.}
}
\description{
Internal function of 
\code{probtrans_by_convolution_clockforward}. It is written in C++ and is
 not meant to be called
directly by the user.
}
\seealso{
\code{\link{probtrans_by_convolution_clockforward}}.
}
\author{
Moritz Gerstung & Rui Costa
}
