% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolute_clockreset}
\alias{convolute_clockreset}
\title{Convolution function for clock-reset models}
\usage{
convolute_clockreset(time_vector, integrand_1, integrand_2)
}
\arguments{
\item{time_vector, integrand_1, integrand_2}{Numeric vectors.}
}
\description{
Internal function of 
\code{probtrans_by_convolution_clockreset}. It is written in C++
and is not meant to be called
directly by the user.
}
\seealso{
\code{\link{probtrans_by_convolution_clockreset}}.
}
\author{
Moritz Gerstung & Rui Costa
}
