% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolute_semiMarkov}
\alias{convolute_semiMarkov}
\title{Convolution function for semi-Markov models}
\usage{
convolute_semiMarkov(time_vector, integrand_1, integrand_2)
}
\arguments{
\item{time_vector, integrand_1, integrand_2}{Numeric vectors.}
}
\description{
Internal function of 
\code{probtrans_by_convolution_semiMarkov}. It is written in C++
and is not meant to be called
directly by the user.
}
\seealso{
\code{\link{probtrans_by_convolution_semiMarkov}}.
}
\author{
Moritz Gerstung & Rui Costa
}
