% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebmstate.R
\name{cumhaz_splines}
\alias{cumhaz_splines}
\title{Spline approximations of the cumulative hazard functions}
\usage{
cumhaz_splines(cumhaz)
}
\arguments{
\item{cumhaz}{An object of class \code{msfit}, created by 
\code{\link{msfit_generic}} or \code{\link{msfit}}.}
}
\value{
A list of estimated cumulative hazard functions (one for each transition).
}
\description{
Creates a spline approximation for the vector of cumulative hazards of each transition.
}
\details{
This function is used by the function \code{probtrans_by_convolution}. It is not meant to be called by the user.
}
\seealso{
\code{\link{msfit_generic}}; \code{\link{msfit}}; \code{\link{probtrans_by_convolution}}.
}
\author{
Rui Costa
}
