% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_and_leave_one_out.R
\name{cumhazCIs_for_target_transition}
\alias{cumhazCIs_for_target_transition}
\title{Ancillary function of \code{boot_ebmstate}.}
\usage{
cumhazCIs_for_target_transition(transition, msfit_objects_boot)
}
\arguments{
\item{transition}{The transition for which transition confidence intervals
are computed.}

\item{msfit_objects_boot}{List of \code{msfit} objects with boostrap estimates 
of cumulative hazard rate functions
for multiple transitions.}
}
\value{
95\% highest density, non-parametric bootstrap confidence intervals for the cumulative
hazard rate functions.
}
\description{
Computes 95\% highest density, non-parametric bootstrap confidence 
intervals for the cumulative hazard rate functions, 
given a list of \code{msfit} objects with boostrap estimates of cumulative hazard rate functions
for multiple transitions. This 
function is not meant to be called by the user.
}
\seealso{
\code{\link{boot_ebmstate}}.
}
\author{
Rui Costa
}
