% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebmstate-package.R
\docType{package}
\name{ebmstate-package}
\alias{ebmstate-package}
\title{Empirical Bayes multi-state Cox model}
\description{
This package implements an empirical Bayes,
multi-state Cox model. Different groups of regression
coefficients can be defined, with coefficients of the
same group sharing the same Gaussian prior. It takes
as input a data set in 'long format' and generates
estimates of relative hazards, cumulative hazard
functions and transition probabilities. It relies 
on packages \code{\link{survival}} and \code{\link{mstate}} 
and incorporates some
of their functions to reduce upstream dependency.
}
\details{
\tabular{ll}{
Package: \tab ebmstate\cr
Type: \tab Package\cr
Version: \tab 0.0.73\cr
Date: \tab 2020-01-21\cr
License: \tab GPL 3\cr
}
}
\author{
Rui Costa, Moritz Gerstung
}
\keyword{package}
