% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{MakeInteger}
\alias{MakeInteger}
\title{Convert factor to integer.}
\usage{
MakeInteger(v)
}
\arguments{
\item{v}{A factor vector.}
}
\value{
A data.frame with columns corresponding to levels in the factor.
}
\description{
Convert factor to integer.
}
\details{
An internal function of \code{CoxRFX}, not meant
to called directly by the user.
}
\seealso{
\code{\link{CoxRFX}}
}
\author{
Moritz Gerstung
}
