% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msfit_object_sample}
\alias{msfit_object_sample}
\title{Estimated cumulative hazard rates under 
an empirical Bayes Cox model (example)}
\format{
An object of class 
\code{c(msfit,coxrfx)}. See \code{\link{msfit_generic}} 
and \code{\link[mstate:msfit]{mstate::msfit}} for details.
}
\usage{
msfit_object_sample
}
\description{
An RData object containing estimated
cumulative hazards,
obtained by running \code{msfit_generic} on the
object \code{coxrfx_object_sample} (also included 
in the present package).
}
\seealso{
\code{\link{coxrfx_object_sample}}.
}
\keyword{datasets}
