% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mstate_data}
\alias{mstate_data}
\title{An example of long-format multistate data}
\format{
A data frame.
}
\usage{
mstate_data
}
\description{
An RData object containing disease progression
data for a sample of 576 patients with myelodysplastic
syndromes (MDS), as an example of long-format 
multistate data.
}
\keyword{datasets}
