% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mstate_data_sample}
\alias{mstate_data_sample}
\title{A simulated event-history data set}
\format{
A data frame with 649 rows and 18 variables
 (250 patients):
\describe{
  \item{id}{patient identification number}
  \item{from}{state in which the patient is}
  \item{to}{state to which the patient is at risk of going to}
  \item{trans}{transition ID number}
  \item{Tstart}{when the risk of the transition started}
  \item{Tstop}{the time at which the risk of the
   transition ended or the last follow-up time (whichever 
   happened first)}
  \item{time}{Tstop-Tstart}
  \item{status}{did the transition occur at Tstop?}
  \item{Cov1,Cov2,Cov3,Cov4,Cov5,Cov6,Cov7,Cov8,Cov9,Cov10}{covariates}
}
}
\usage{
mstate_data_sample
}
\description{
A data set generated by simulation from an 
illness-death Cox model. This is an object
of double class 'data.frame' and 'msdata', whose
'trans' attribute is a transition matrix  
(\code{attr(mstate_data_sample,"trans")}).
}
\keyword{datasets}
