% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{print.msfit}
\alias{print.msfit}
\title{Print method for \code{msfit} objects
 generated by \code{msfit_generic}}
\usage{
\method{print}{msfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{msfit} or double class \code{msfit} 
and \code{coxrfx}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This method is a simple call to \code{print.default}.
Its main purpose is to override \code{print.coxrfx}
when printing an
object of double class \code{msfit} and \code{coxrfx}.
}
\author{
Rui Costa
}
