% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{print.coxrfx}
\alias{print.coxrfx}
\title{Print method for CoxRFX objects}
\usage{
\method{print}{coxrfx}(x, ...)
}
\arguments{
\item{x}{A \code{coxrfx} object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns an invisible NULL object.
}
\description{
This function implicitly calls summary.coxrfx().
}
\details{
Prints two data frames, one with hyperparameter estimates and 
another with regression coefficient estimates.
}
\author{
Moritz Gerstung & Rui Costa
}
